Fixed bug where adding a new status effect in battle was causing a crash (was trying to access the newly added member from the list of previously added members)
So it now checks 3 cases, does this unit already have a status effect of the same type?  If so, just refresh the effect.
Is this status effect being loaded in from a previous state/scene?  If so, carry over the ticks that happened in that scene.
If neither of the above are true, this must be a brand new fresh effect being applied to a unit that doesn't have an effect of that type already, so add a fresh'un.